refactor(svelte2tsx): extract AST name helpers into script/ast_utils.rs
Pure move: no logic changes, only `use` additions and `pub(super)` visibility
adjustments. The corresponding unit tests move with the code.
