+++ crates/rsvelte_core/src/svelte2tsx/script/mod.rs
@@ -8,6 +8,7 @@
 //! the OXC AST directly. This avoids dependency on the thread-local ParseArena
 //! used by the main compiler, keeping svelte2tsx self-contained.
 
+mod ast_utils;
 mod component_events;
 mod hoistable_types;
 mod parse;
@@ -31,6 +32,12 @@ use super::svelte2tsx::slice_src;
 
 pub use component_events::{ComponentEvents, EventInfo};
 
+use ast_utils::{
+    binding_pattern_simple_name, collect_binding_names, collect_top_level_declared_names,
+    declarator_has_boolean_init, extract_all_names_from_binding_pattern,
+    extract_names_from_assignment_target, extract_names_from_binding_pattern_full,
+    module_export_name_to_string, property_key_to_string,
+};
 use hoistable_types::{
     HoistCandidate, hoist_dollar_generic_referenced_types, is_ident_char_for_str,
     is_special_type_name, resolve_hoistable_type_decls, rewrite_interface_to_type_dts,
@@ -3470,152 +3477,6 @@ fn find_matching_brace(text: &str) -> Option<usize> {
     None
 }
 
-fn extract_names_from_binding_pattern_full(
-    pattern: &oxc::BindingPattern,
-    exported_names: &mut ExportedNames,
-    has_default: bool,
-    is_prop: bool,
-    is_let: bool,
-    is_named_export: bool,
-) {
-    match pattern {
-        oxc::BindingPattern::BindingIdentifier(id) => {
-            let name = id.name.to_string();
-            exported_names.add_full(
-                name.clone(),
-                name,
-                has_default,
-                None,
-                is_prop,
-                is_let,
-                is_named_export,
-            );
-        }
-        oxc::BindingPattern::ObjectPattern(obj_pat) => {
-            for prop in obj_pat.properties.iter() {
-                match &prop.value {
-                    oxc::BindingPattern::AssignmentPattern(assign) => {
-                        extract_names_from_binding_pattern_full(
-                            &assign.left,
-                            exported_names,
-                            true,
-                            is_prop,
-                            is_let,
-                            is_named_export,
-                        );
-                    }
-                    _ => {
-                        extract_names_from_binding_pattern_full(
-                            &prop.value,
-                            exported_names,
-                            has_default,
-                            is_prop,
-                            is_let,
-                            is_named_export,
-                        );
-                    }
-                }
-            }
-            // Handle rest element: `{ a, ...rest }` — recurse into `rest`
-            if let Some(rest) = &obj_pat.rest {
-                extract_names_from_binding_pattern_full(
-                    &rest.argument,
-                    exported_names,
-                    has_default,
-                    is_prop,
-                    is_let,
-                    is_named_export,
-                );
-            }
-        }
-        oxc::BindingPattern::ArrayPattern(arr_pat) => {
-            for el in arr_pat.elements.iter().flatten() {
-                match el {
-                    oxc::BindingPattern::AssignmentPattern(assign) => {
-                        extract_names_from_binding_pattern_full(
-                            &assign.left,
-                            exported_names,
-                            true,
-                            is_prop,
-                            is_let,
-                            is_named_export,
-                        );
-                    }
-                    _ => {
-                        extract_names_from_binding_pattern_full(
-                            el,
-                            exported_names,
-                            has_default,
-                            is_prop,
-                            is_let,
-                            is_named_export,
-                        );
-                    }
-                }
-            }
-            // Handle rest element: `[a, ...rest]` — recurse into `rest`
-            if let Some(rest) = &arr_pat.rest {
-                extract_names_from_binding_pattern_full(
-                    &rest.argument,
-                    exported_names,
-                    has_default,
-                    is_prop,
-                    is_let,
-                    is_named_export,
-                );
-            }
-        }
-        oxc::BindingPattern::AssignmentPattern(assign) => {
-            extract_names_from_binding_pattern_full(
-                &assign.left,
-                exported_names,
-                true,
-                is_prop,
-                is_let,
-                is_named_export,
-            );
-        }
-    }
-}
-
-/// Get a simple name from a binding pattern (only works for BindingIdentifier).
-fn binding_pattern_simple_name(pattern: &oxc::BindingPattern) -> Option<String> {
-    match pattern {
-        oxc::BindingPattern::BindingIdentifier(id) => Some(id.name.to_string()),
-        _ => None,
-    }
-}
-
-/// Whether a declarator's initializer is a boolean literal (`let x = false`).
-/// Mirrors official `propTypeAssertToUserDefined`'s `True/FalseKeyword` check —
-/// such an init still forces the `__sveltets_2_any` widen (TS would otherwise
-/// narrow `x` to the `false`/`true` literal type).
-fn declarator_has_boolean_init(declarator: &oxc::VariableDeclarator) -> bool {
-    declarator
-        .init
-        .as_ref()
-        .is_some_and(|init| matches!(init, oxc::Expression::BooleanLiteral(_)))
-}
-
-/// Convert a PropertyKey to a string name.
-fn property_key_to_string(key: &oxc::PropertyKey) -> Option<String> {
-    match key {
-        oxc::PropertyKey::StaticIdentifier(id) => Some(id.name.to_string()),
-        oxc::PropertyKey::StringLiteral(lit) => Some(lit.value.to_string()),
-        oxc::PropertyKey::NumericLiteral(lit) => Some(lit.value.to_string()),
-        _ => None,
-    }
-}
-
-/// Convert a ModuleExportName to a string.
-fn module_export_name_to_string(name: &oxc::ModuleExportName) -> String {
-    match name {
-        oxc::ModuleExportName::IdentifierName(id) => id.name.to_string(),
-        oxc::ModuleExportName::IdentifierReference(id) => id.name.to_string(),
-        oxc::ModuleExportName::StringLiteral(lit) => lit.value.to_string(),
-    }
-}
-
 // =============================================================================
 // Store subscription injection
 // =============================================================================
@@ -4184,193 +4045,6 @@ impl<'a> Visit<'a> for DollarParamShadowCollector {
     fn visit_arrow_function_expression(&mut self, it: &oxc::ArrowFunctionExpression<'a>) {
         self.add_params(&it.params, it.span);
         oxc_ast_visit::walk::walk_arrow_function_expression(self, it);
-    }
-}
-
-/// Pre-pass: collect EVERY top-level declared binding name in the instance
-/// script before rune detection runs. Official `svelte2tsx` resolves a
-/// `$name` reference as a store auto-subscription (NOT the `$state`/`$derived`/
-/// `$effect` rune) whenever `name` is a declared binding, using the COMPLETE
-/// top-level scope. So `let state = $state(0)` must see its own `state` as
-/// declared (→ legacy), while `let x = $state(0)` stays runes. Without this
-/// pre-pass `declared_names` was still empty when a declarator's own
-/// initializer was checked, over-detecting runes. Mirrors upstream
-/// `ImplicitStoreValues` / `checkGlobalsForRunes`.
-fn collect_top_level_declared_names(body: &[oxc::Statement]) -> HashSet<String> {
-    fn add_var(vd: &oxc::VariableDeclaration, names: &mut HashSet<String>) {
-        for d in vd.declarations.iter() {
-            for n in extract_all_names_from_binding_pattern(&d.id) {
-                names.insert(n);
-            }
-        }
-    }
-    let mut names = HashSet::new();
-    for stmt in body {
-        match stmt {
-            oxc::Statement::VariableDeclaration(vd) => add_var(vd, &mut names),
-            oxc::Statement::FunctionDeclaration(f) => {
-                if let Some(id) = &f.id {
-                    names.insert(id.name.to_string());
-                }
-            }
-            oxc::Statement::ClassDeclaration(c) => {
-                if let Some(id) = &c.id {
-                    names.insert(id.name.to_string());
-                }
-            }
-            oxc::Statement::TSModuleDeclaration(m) => {
-                if let oxc_ast::ast::TSModuleDeclarationName::Identifier(id) = &m.id {
-                    names.insert(id.name.to_string());
-                }
-            }
-            oxc::Statement::TSEnumDeclaration(e) => {
-                names.insert(e.id.name.to_string());
-            }
-            oxc::Statement::ImportDeclaration(imp) => {
-                if let Some(specs) = &imp.specifiers {
-                    for s in specs.iter() {
-                        let n = match s {
-                            oxc::ImportDeclarationSpecifier::ImportDefaultSpecifier(s) => {
-                                s.local.name.to_string()
-                            }
-                            oxc::ImportDeclarationSpecifier::ImportNamespaceSpecifier(s) => {
-                                s.local.name.to_string()
-                            }
-                            oxc::ImportDeclarationSpecifier::ImportSpecifier(s) => {
-                                s.local.name.to_string()
-                            }
-                        };
-                        names.insert(n);
-                    }
-                }
-            }
-            oxc::Statement::ExportNamedDeclaration(ex) => {
-                if let Some(decl) = &ex.declaration {
-                    match decl {
-                        oxc::Declaration::VariableDeclaration(vd) => add_var(vd, &mut names),
-                        oxc::Declaration::FunctionDeclaration(f) => {
-                            if let Some(id) = &f.id {
-                                names.insert(id.name.to_string());
-                            }
-                        }
-                        oxc::Declaration::ClassDeclaration(c) => {
-                            if let Some(id) = &c.id {
-                                names.insert(id.name.to_string());
-                            }
-                        }
-                        _ => {}
-                    }
-                }
-            }
-            _ => {}
-        }
-    }
-    names
-}
-
-/// Extract all identifier names from a binding pattern (for destructuring support).
-///
-/// For `{ a, b, c }` returns `["a", "b", "c"]`.
-/// For `[a, b, c]` returns `["a", "b", "c"]`.
-/// For simple identifiers, returns the single name.
-fn extract_all_names_from_binding_pattern(pattern: &oxc::BindingPattern) -> Vec<String> {
-    let mut names = Vec::new();
-    collect_binding_names(pattern, &mut names);
-    names
-}
-
-fn collect_binding_names(pattern: &oxc::BindingPattern, names: &mut Vec<String>) {
-    match pattern {
-        oxc::BindingPattern::BindingIdentifier(id) => {
-            names.push(id.name.to_string());
-        }
-        oxc::BindingPattern::ObjectPattern(obj) => {
-            for prop in obj.properties.iter() {
-                collect_binding_names(&prop.value, names);
-            }
-            if let Some(ref rest) = obj.rest {
-                collect_binding_names(&rest.argument, names);
-            }
-        }
-        oxc::BindingPattern::ArrayPattern(arr) => {
-            for el in arr.elements.iter().flatten() {
-                collect_binding_names(el, names);
-            }
-            if let Some(ref rest) = arr.rest {
-                collect_binding_names(&rest.argument, names);
-            }
-        }
-        oxc::BindingPattern::AssignmentPattern(assign) => {
-            collect_binding_names(&assign.left, names);
-        }
-    }
-}
-
-/// Extract names from the left-hand side of an assignment expression
-/// (used for reactive declarations like `$: store = ...`).
-fn extract_names_from_assignment_target(target: &oxc::AssignmentTarget) -> Vec<String> {
-    let mut names = Vec::new();
-    collect_assignment_target_names(target, &mut names);
-    names
-}
-
-fn collect_assignment_target_names(target: &oxc::AssignmentTarget, names: &mut Vec<String>) {
-    match target {
-        oxc::AssignmentTarget::AssignmentTargetIdentifier(id) => {
-            let name = id.name.to_string();
-            if !name.starts_with('$') {
-                names.push(name);
-            }
-        }
-        oxc::AssignmentTarget::ObjectAssignmentTarget(obj) => {
-            for prop in obj.properties.iter() {
-                match prop {
-                    oxc::AssignmentTargetProperty::AssignmentTargetPropertyIdentifier(id) => {
-                        let name = id.binding.name.to_string();
-                        if !name.starts_with('$') {
-                            names.push(name);
-                        }
-                    }
-                    oxc::AssignmentTargetProperty::AssignmentTargetPropertyProperty(prop) => {
-                        match &prop.binding {
-                            oxc::AssignmentTargetMaybeDefault::AssignmentTargetWithDefault(
-                                with_default,
-                            ) => {
-                                collect_assignment_target_names(&with_default.binding, names);
-                            }
-                            _ => {
-                                if let Some(target) = prop.binding.as_assignment_target() {
-                                    collect_assignment_target_names(target, names);
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-            if let Some(ref rest) = obj.rest {
-                collect_assignment_target_names(&rest.target, names);
-            }
-        }
-        oxc::AssignmentTarget::ArrayAssignmentTarget(arr) => {
-            for el in arr.elements.iter().flatten() {
-                match el {
-                    oxc::AssignmentTargetMaybeDefault::AssignmentTargetWithDefault(
-                        with_default,
-                    ) => {
-                        collect_assignment_target_names(&with_default.binding, names);
-                    }
-                    _ => {
-                        if let Some(target) = el.as_assignment_target() {
-                            collect_assignment_target_names(target, names);
-                        }
-                    }
-                }
-            }
-            if let Some(ref rest) = arr.rest {
-                collect_assignment_target_names(&rest.target, names);
-            }
-        }
-        _ => {}
     }
 }
